+++ tests/Integration/Internal/ComposerPlugin/ComponentInstallerTest.php
@@ -24,7 +24,7 @@ class ComponentInstallerTest extends TestCase
 {
     use ContainerTrait;
 
-    private $servicesFilePath = __DIR__ . '/Fixtures/services.yaml';
+    private $servicesFilePath = 'Fixtures/services.yaml';
 
     public function tearDown(): void
     {