+++ tests/Integration/Internal/Framework/Module/Setup/Service/ModuleServicesActivationServiceTest.php
@@ -24,6 +24,7 @@ use OxidEsales\EshopCommunity\Tests\Unit\Internal\ContextStub;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Webmozart\PathUtil\Path;
 
 class ModuleServicesActivationServiceTest extends TestCase
 {
@@ -117,7 +118,7 @@ class ModuleServicesActivationServiceTest extends TestCase
         $this->shopActivationService->activateModuleServices($this->testModuleId, 4);
         $this->shopActivationService->activateModuleServices($this->testModuleId, 5);
 
-        $this->assertProjectYamlHasImport($this->getTestModuleServiceYamlPath());
+        $this->assertProjectYamlHasImport($this->getRelativeTestServiceYamlPath());
         $this->assertModuleServiceIsActiveForShops('testEventSubscriber', [1,4,5]);
     }
 
@@ -170,7 +171,7 @@ class ModuleServicesActivationServiceTest extends TestCase
 
         $projectConfig = new DIConfigWrapper([
             'imports' => [
-                ['resource' => $this->getTestModuleServiceYamlPath()]
+                ['resource' => $this->getRelativeTestServiceYamlPath()]
             ],
             'services' => [
                 'testEventSubscriber' => [
@@ -254,5 +255,13 @@ class ModuleServicesActivationServiceTest extends TestCase
     private function getTestModuleServiceYamlPath(): string
     {
         return realpath($this->testModuleDirectory . DIRECTORY_SEPARATOR . 'services.yaml');
+    }
+
+    public function getRelativeTestServiceYamlPath(): string
+    {
+        return Path::makeRelative(
+            $this->getTestModuleServiceYamlPath(),
+            Path::getDirectory($this->contextStub->getGeneratedServicesFilePath())
+        );
     }
 }